Manual weight set (#62)
* manual weight set

* pinning numpy version (#59)
