+++ autoupdate_validator_steps.sh
@@ -21,5 +21,5 @@
 #!/bin/bash
 
 # Function to run a command with sudo only if it is available
-./launch_validators.sh
+# ./launch_validators.sh
 echo "Autoupdate steps complete :)"
