+++ requirements.txt
@@ -10,3 +10,4 @@ typer==0.10.0
 xgboost==2.0.3
 scikit-learn
 sqlite-web==0.6.3
+numpy==1.26.4
